requirements.txt
fastapi==0.115.12
groq==0.22.0
langchain==0.3.23
langchain_core==0.3.54
langchain_groq==0.3.2
mistralai==1.7.0
Pillow==9.5.0
pydantic==2.11.3
python-dotenv==1.1.0
sentence_transformers==3.4.1
torch==2.0.1
torchvision==0.15.2
transformers==4.49.0
uvicorn==0.34.1
